// frontend/src/model/updateUserPreferencesRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * cinematch-api
 * OpenAPI spec version: 0.1.0
 */

/**
 * User-facing update struct: genre names instead of UUIDs
 */
export interface UpdateUserPreferencesRequest {
  /** @nullable */
  exclude_genres?: string[] | null;
  /** @nullable */
  include_genres?: string[] | null;
  /** @nullable */
  is_tite?: boolean | null;
  /** @nullable */
  release_year_flex?: number | null;
  /** @nullable */
  target_release_year?: number | null;
}